update wmma check
