--- include/ck_tile/core/arch/wmma_config.hpp
@@ -1,52 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Copyright (c) 2025, Advanced Micro Devices, Inc. All rights reserved.
-
-#pragma once
-
-#include "ck_tile/core.hpp"
-#include "ck_tile/host/device_prop.hpp"
-
-namespace ck_tile {
-// Architecture tags
-struct gfx11_t
-{
-};
-struct gfx12_t
-{
-};
-
-template <typename ADataType,
-          typename BDataType,
-          typename AccDataType,
-          index_t M_Warp_Tile,
-          index_t N_Warp_Tile,
-          index_t K_Warp_Tile>
-CK_TILE_HOST bool check_wmma_supported()
-{
-    if(is_gfx12_supported())
-    {
-        return has_wmma_traits_v<gfx12_t,
-                                 ADataType,
-                                 BDataType,
-                                 AccDataType,
-                                 M_Warp_Tile,
-                                 N_Warp_Tile,
-                                 K_Warp_Tile>;
-    }
-    else if(is_gfx11_supported())
-    {
-        return has_wmma_traits_v<gfx11_t,
-                                 ADataType,
-                                 BDataType,
-                                 AccDataType,
-                                 M_Warp_Tile,
-                                 N_Warp_Tile,
-                                 K_Warp_Tile>;
-    }
-    else
-    {
-        return false;
-    }
-}
-
-} // namespace ck_tile
